+++ src/commands/servers/koa-next.ts
@@ -1,3 +1,4 @@
+// !deprecated
 import path from 'path';
 import http from 'http';
 import https from 'https';